% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add}
\alias{add}
\title{Function adding up two floats}
\usage{
add(firstOperand, secondOperand)
}
\arguments{
\item{firstOperand}{as first operand}

\item{secondOperand}{as second operand}
}
\description{
Function adding up two floats
}
