% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{divide}
\alias{divide}
\title{Function dividing two floats}
\usage{
divide(firstOperand, secondOperand)
}
\arguments{
\item{firstOperand}{as first operand}

\item{secondOperand}{as second operand}
}
\description{
Function dividing two floats
}
