% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply}
\alias{multiply}
\title{Function multiplying two floats}
\usage{
multiply(firstOperand, secondOperand)
}
\arguments{
\item{firstOperand}{as first operand}

\item{secondOperand}{as second operand}
}
\description{
Function multiplying two floats
}
