% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subtract}
\alias{subtract}
\title{Function subtracting two floats}
\usage{
subtract(firstOperand, secondOperand)
}
\arguments{
\item{firstOperand}{as first operand}

\item{secondOperand}{as second operand}
}
\description{
Function subtracting two floats
}
