% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculator.R
\name{SubtractOperation}
\alias{SubtractOperation}
\title{SubtractOperation}
\usage{
SubtractOperation(a, b)
}
\arguments{
\item{a}{as first operand}

\item{b}{as second operand}
}
\description{
SubtractOperation
}
